requirements.txt
azure-storage-blob==12.17.0
beautifulsoup4==4.12.2
numpy==1.25.2
pandas==2.0.3
psycopg2==2.9.7
pyarrow==12.0.1
Requests==2.31.0
SQLAlchemy==2.0.20
python-dotenv==1.0.0
cloudscraper==1.2.71
lxml==5.4.0
